requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
python-dotenv
alembic
# auth / JWT
python-jose[cryptography]
bcrypt
python-multipart
# testing
pytest
httpx
# document parsing
pypdf
python-docx
openpyxl
xlrd
